# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PautaPro API",
    "description": "API Empresarial para Integración de Sistemas de Gestión Publicitaria",
    "version": "2.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@pautapro.com",
      "url": "https://developers.pautapro.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.pautapro.com/v2",
      "description": "Servidor de Producción"
    },
    {
      "url": "https://api-staging.pautapro.com/v2",
      "description": "Servidor de Staging"
    },
    {
      "url": "http://localhost:3001/api/v2",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para autenticación empresarial"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT para usuarios autenticados"
      }
    },
    "schemas": {
      "Cliente": {
        "type": "object",
        "required": ["nombrecliente", "rut", "email"],
        "properties": {
          "id_cliente": {
            "type": "integer",
            "description": "ID único del cliente",
            "example": 12345
          },
          "nombrecliente": {
            "type": "string",
            "description": "Nombre del cliente",
            "example": "Empresa ABC SpA"
          },
          "nombrefantasia": {
            "type": "string",
            "description": "Nombre de fantasía",
            "example": "ABC Corp"
          },
          "razonsocial": {
            "type": "string",
            "description": "Razón social",
            "example": "Empresa ABC Sociedad por Acciones"
          },
          "rut": {
            "type": "string",
            "description": "RUT del cliente",
            "example": "12.345.678-9"
          },
          "id_region": {
            "type": "integer",
            "description": "ID de la región",
            "example": 7
          },
          "id_comuna": {
            "type": "integer",
            "description": "ID de la comuna",
            "example": 123
          },
          "direccion": {
            "type": "string",
            "description": "Dirección del cliente",
            "example": "Av. Providencia 1234, Santiago"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del cliente",
            "example": "contacto@empresaabc.com"
          },
          "telcelular": {
            "type": "string",
            "description": "Teléfono celular",
            "example": "+56912345678"
          },
          "giro": {
            "type": "string",
            "description": "Giro comercial",
            "example": "Servicios de consultoría"
          },
          "id_grupo": {
            "type": "integer",
            "description": "ID del grupo empresarial",
            "example": 10
          },
          "id_tipo_cliente": {
            "type": "integer",
            "description": "ID del tipo de cliente",
            "example": 1
          },
          "nombrerepresentantelegal": {
            "type": "string",
            "description": "Nombre del representante legal",
            "example": "Juan"
          },
          "rut_representante": {
            "type": "string",
            "description": "RUT del representante legal",
            "example": "98.765.432-1"
          },
          "estado": {
            "type": "boolean",
            "description": "Estado del cliente",
            "example": true
          },
          "fechacreacion": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación",
            "example": "2024-11-13T17:29:40.340Z"
          }
        }
      },
...
    }
  },
  "paths": {
    "/clientes": {
      "get": {
        "summary": "Listar clientes",
        "description": "Obtiene una lista paginada de clientes con filtros avanzados",
        "tags": ["Clientes"],
        "security": ["ApiKeyAuth"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Límite de elementos por página",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Término de búsqueda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "description": "Filtrar por estado",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Cliente"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de requests excedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear nuevo cliente",
        "description": "Crea un nuevo cliente en el sistema",
        "tags": ["Clientes"],
        "security": ["ApiKeyAuth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombrecliente", "rut", "email"],
                "properties": {
                  "nombrecliente": {
                    "type": "string"
                  },
                  "rut": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "nombrefantasia": {
                    "type": "string"
                  },
                  "razonsocial": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "giro": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cliente creado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cliente"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "RUT duplicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "summary": "Obtener cliente específico",
        "description": "Obtiene los detalles de un cliente específico",
        "tags": ["Clientes"],
        "security": ["ApiKeyAuth"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del cliente",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cliente"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Actualizar cliente",
        "description": "Actualiza los datos de un cliente específico",
        "tags": ["Clientes"],
        "security": ["ApiKeyAuth"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del cliente",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombrecliente": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "direccion": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cliente"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Eliminar cliente",
        "description": "Elimina un cliente del sistema (soft delete)",
        "tags": ["Clientes"],
        "security": ["ApiKeyAuth"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del cliente",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cliente eliminado exitosamente"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "No se puede eliminar cliente con órdenes activas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}